backend/requirements.txt
fastapi
uvicorn
python-multipart
firebase-admin
python-dotenv
pydantic
passlib[bcrypt]
python-jose
requests
git+https://github.com/scaledown-team/scaledown.git
tiktoken
groq